third_party/blink/web_tests/platform/generic/fast/dom/MutationObserver/removed-out-of-order-expected.txt

Test MutationEvents interfering with MutationObservers: removing nodes 'out of order'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 2
PASS mutations[0].removedNodes.length is 0
PASS mutations[0].addedNodes[0].tagName is 'B'
PASS mutations[1].addedNodes.length is 0
PASS mutations[1].removedNodes.length is 1
PASS mutations[1].removedNodes[0].tagName is 'B'
PASS mutations[1].addedNodes[0] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
